Change: Files For Connection And Relationship

+++ src/main/java/com/uisrael/drinkhouse/infraestructura/repositorio/IProductoJpaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.uisrael.drinkhouse.infraestructura.persistencia.jpa.ProductoEntity;
 
-public interface IProductoJpaRepositorio extends  JpaRepository<ProductoEntity, Integer>{
+public interface IProductoJpaRepositorio extends JpaRepository<ProductoEntity, Long> {
 
 }
